Move some classes around, added force save command
