+++ src/main/java/goose/politik/commands/ListCommand.java
@@ -1,7 +1,7 @@
 package goose.politik.commands;
 
 import goose.politik.Politik;
-import goose.politik.util.player.PolitikPlayer;
+import goose.politik.player.PolitikPlayer;
 import goose.politik.util.landUtil.LandUtil;
 import org.bukkit.World;
 import org.bukkit.command.Command;
