+++ src/main/java/goose/politik/commands/NationCommands.java
@@ -1,8 +1,8 @@
 package goose.politik.commands;
 
-import goose.politik.util.config.ConfigHandler;
-import goose.politik.util.government.nation.Nation;
-import goose.politik.util.player.PolitikPlayer;
+import goose.politik.config.ConfigHandler;
+import goose.politik.government.nation.Nation;
+import goose.politik.player.PolitikPlayer;
 import goose.politik.util.text.TextUtil;
 import net.kyori.adventure.text.Component;
 import org.bukkit.command.Command;
