+++ src/main/java/goose/politik/commands/PolitikSaveCommand.java
@@ -1,5 +1,6 @@
 package goose.politik.commands;
 
+import goose.politik.Politik;
 import goose.politik.player.PolitikPlayer;
 import goose.politik.util.text.TextUtil;
 import org.bukkit.command.Command;
@@ -8,17 +9,13 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 
-public class BalCommand implements CommandExecutor {
-
+public class PolitikSaveCommand implements CommandExecutor {
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, String[] args) {
-        if (sender instanceof Player) {
-            PolitikPlayer player = PolitikPlayer.getPolitikPlayer(((Player) sender));
-            //player not console
-            player.message(TextUtil.successMessage("Your balance is: $" + player.getMoney()));
-        } else {
-            sender.sendMessage(TextUtil.errorMessage("Do you really need this?"));
+        if (!sender.isOp()) {
+            return true;
         }
+        Politik.getInstance().saveServer();
         return true;
     }
 }
