+++ src/main/java/goose/politik/commands/SetJobCommand.java
@@ -1,7 +1,7 @@
 package goose.politik.commands;
 
 import goose.politik.Politik;
-import goose.politik.util.player.PolitikPlayer;
+import goose.politik.player.PolitikPlayer;
 import goose.politik.util.text.TextUtil;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
