+++ src/main/java/goose/politik/commands/SetMoneyCommand.java
@@ -2,7 +2,7 @@ package goose.politik.commands;
 
 import goose.politik.Politik;
 import goose.politik.util.economy.MoneyHandler;
-import goose.politik.util.player.PolitikPlayer;
+import goose.politik.player.PolitikPlayer;
 import goose.politik.util.text.TextUtil;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
