+++ src/main/java/goose/politik/government/town/TownConfig.java
@@ -1,0 +1,4 @@
+package goose.politik.government.town;
+
+public class TownConfig {
+}
