+++ src/main/java/goose/politik/util/database/NationDB.java
@@ -1,8 +1,8 @@
 package goose.politik.util.database;
 
-import goose.politik.util.government.nation.Nation;
-import goose.politik.util.player.PolitikPlayer;
-import goose.politik.util.government.town.Town;
+import goose.politik.government.nation.Nation;
+import goose.politik.player.PolitikPlayer;
+import goose.politik.government.town.Town;
 import org.bson.Document;
 
 import java.math.BigDecimal;
