--- src/main/java/goose/politik/util/government/town/TownConfig.java
@@ -1,4 +1,0 @@
-package goose.politik.util.government.town;
-
-public class TownConfig {
-}
